use for, (void) and std::as_const
https://godbolt.org/z/W1oM15Evj
I just made some test code things appear to still work.

For loop can iterate and check the condition in the same line.

https://blog.codeisc.com/2018/01/09/cpp-comma-operator-nice-usages.html
https://blog.codeisc.com/2017/12/26/cpp-comma-operator-introduction.html
https://stackoverflow.com/questions/39514765/the-void-the-comma-operator-operator-and-the-impossible-overloading
https://youtu.be/AqDsso3S5fg?t=276
(void), comma operator can be overloaded. So you need to cast to void to force using the built-in comma operator.

https://stackoverflow.com/questions/5853833/casting-non-const-to-const-in-c
https://en.cppreference.com/w/cpp/utility/as_const
I was worried that a pred could mutate the values as there was nothing to prevent this. As the Iterator needs to be mutable so you can iterate. The thing it points too may or may not be const. So someone could write a pred that mutates. Same with Op for transform. You don't want to mutate the incoming just the out going value.